Rewrite Init, Read, Write, Destroy commands and their specs

+++ app/commands/repository/filesystem/destroy.rb
@@ -7,6 +7,8 @@ module Repository
     #
     class Destroy < YAMLCommand
       def execute
+        raise OpenWebslides::RepoDoesNotExistError unless Dir.exist? repo_path
+
         FileUtils.rm_r repo_path, :secure => true
       end
     end
